Fixed warnings caused by type conversions

+++ GridsGallery/GridsGallery/GridsGalleryCommonViewController.h
@@ -32,7 +32,7 @@
 - (void)addColumns;
 
 #pragma mark - Utility methods
-- (NSString*)getGrossStringForValue:(int)value;
+- (NSString*)getGrossStringForValue:(NSInteger)value;
 - (void)addColumnWithTitle:(NSString*)title propertyKey:(NSString*)propertyKey
                      width:(float)width textAlignment:(NSTextAlignment)textAlignment
                   cellType:(Class)cellType edgeInsets:(UIEdgeInsets)edgeInsets;
